+++ src/ErrorBoundary/ErrorBoundary.tsx
@@ -4,20 +4,15 @@
  * SPDX-License-Identifier: LicenseRef-Nordic-4-Clause
  */
 
-/* eslint-disable @typescript-eslint/ban-ts-comment */
-
 import React, { ReactNode } from 'react';
 import { Button } from 'react-bootstrap';
 import { connect } from 'react-redux';
-import PropTypes from 'prop-types';
+import { Device } from 'pc-nrfconnect-shared';
 
-import deviceShape, {
-    DeviceShapeProps,
-} from '../Device/DeviceSelector/deviceShape';
 import Spinner from '../Dialog/Spinner';
 import FactoryResetButton from '../FactoryReset/FactoryResetButton';
 import { CollapsibleGroup } from '../SidePanel/Group';
-import { RootState } from '../state';
+import { Devices, RootState } from '../state';
 import { openUrl } from '../utils/open';
 import packageJson from '../utils/packageJson';
 import { getAppSpecificStore as store } from '../utils/persistentStore';
@@ -47,8 +42,8 @@ const sendGAEvent = (error: string) => {
 
 interface Props {
     children: ReactNode;
-    selectedSerialNumber?: string | null;
-    devices?: DeviceShapeProps;
+    selectedSerialNumber?: string;
+    devices: Devices;
     appName?: string;
     restoreDefaults?: () => void;
     sendUsageData?: (message: string) => void;
@@ -80,14 +75,16 @@ class ErrorBoundary extends React.Component<
             ? sendUsageData(error.message)
             : sendGAEvent(error.message);
 
+        const selectedDevice =
+            selectedSerialNumber == null
+                ? undefined
+                : devices[selectedSerialNumber];
+
         generateSystemReport(
             new Date().toISOString().replace(/:/g, '-'),
-            // @ts-ignore Don't know what to do with these yet
-            Object.values(devices),
-            // @ts-ignore Don't know what to do with these yet
-            selectedSerialNumber,
-            // @ts-ignore Don't know what to do with these yet
-            devices[selectedSerialNumber]
+            Object.values(devices) as Device[],
+            selectedDevice,
+            selectedSerialNumber
         ).then(report => {
             this.setState({ systemReport: report });
         });
@@ -181,7 +178,6 @@ class ErrorBoundary extends React.Component<
                                 'https://devzone.nordicsemi.com/support/add'
                             )
                         }
-                        disabled={!systemReport}
                     >
                         Go to DevZone
                     </Button>
@@ -191,20 +187,9 @@ class ErrorBoundary extends React.Component<
     }
 }
 
-// @ts-ignore Don't know what to do with these yet
-ErrorBoundary.propTypes = {
-    children: PropTypes.node.isRequired,
-    selectedSerialNumber: PropTypes.string,
-    devices: PropTypes.objectOf(deviceShape),
-    appName: PropTypes.string,
-    restoreDefaults: PropTypes.func,
-    sendUsageData: PropTypes.func,
-};
-
 const mapStateToProps = (state: RootState) => ({
-    devices: state.device?.devices || {},
-    selectedSerialNumber: state.device?.selectedSerialNumber,
+    devices: state.device?.devices ?? {},
+    selectedSerialNumber: state.device?.selectedSerialNumber ?? undefined,
 });
 
-// @ts-ignore Don't know what to do with these yet
 export default connect(mapStateToProps)(ErrorBoundary);